requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.7
idna==2.8
importlib-metadata==0.23
mccabe==0.6.1
more-itertools==7.2.0
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-gitlab==1.11.0
requests==2.22.0
six==1.12.0
toml==0.10.0
tox==3.9.0
urllib3==1.25.5
virtualenv==16.7.5
zipp==0.6.0
